Add files via upload
This code defines a function flip_coin that simulates the flipping of a coin by returning 'heads' or 'tails' randomly. It also defines a function simulate_coin_flips that simulates the flipping of a coin a specified number of times and returns a list of the results. Finally, it defines a function count_heads that counts the number of heads in a list of coin flips.

The code then uses these functions to simulate the flipping of a coin 1000 times and calculates the probability of heads. It then simulates the flipping of a coin 100 more times and calculates the probability of heads again.

Running this code should demonstrate that the probability of heads or tails does not change based on the outcome of previous flips.